requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Werkzeug>=2.0
pymysql>=1.0
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
matplotlib>=3.4
folium>=0.14.0
geopy>=2.3.0